Spacing is CGFloat

+++ Sources/DesignSystem/SpacingManager.swift
@@ -8,13 +8,13 @@
 import Foundation
 import UIKit
 
-public enum Spacing: Float {
+public enum Spacing: CGFloat {
     case small = 4
     case medium = 8
     case large = 16
     case superLarge = 24
 
-    public var name: Float {
+    public var name: CGFloat {
         return self.rawValue
     }
 }
